elevaite/python_apps/agent_studio/pyproject.toml
[project]
name = "agent-studio"
version = "0.1.0"
description = "Agent Studio"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.111.0",
    "httpx>=0.27.0",
    "markdownify>=1.1.0",
    "openai>=1.77.0",
    "pydantic>=2.7.1",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "uvicorn>=0.34.2",
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "pika>=1.3.2",
    "redis>=4.5.1",
    "elevaite-logger",
    "pytest-asyncio>=0.26.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "pytest-cov>=4.1.0"]

[tool.uv.sources]
elevaite-logger = { workspace = true }
